Move decoder tests to a separate file
