+++ library/src/test/java/com/chuckerteam/chucker/api/ChuckerInterceptorTest.kt
@@ -1,22 +1,21 @@
 package com.chuckerteam.chucker.api
 
-import com.chuckerteam.chucker.ChuckerInterceptorDelegate
-import com.chuckerteam.chucker.NoLoggerRule
-import com.chuckerteam.chucker.SEGMENT_SIZE
-import com.chuckerteam.chucker.getResourceFile
-import com.chuckerteam.chucker.readByteStringBody
+import com.chuckerteam.chucker.util.ChuckerInterceptorDelegate
+import com.chuckerteam.chucker.util.ClientFactory
+import com.chuckerteam.chucker.util.NoLoggerRule
+import com.chuckerteam.chucker.util.SEGMENT_SIZE
+import com.chuckerteam.chucker.util.getResourceFile
+import com.chuckerteam.chucker.util.readByteStringBody
+import com.chuckerteam.chucker.util.toServerRequest
 import com.google.common.collect.Range
 import com.google.common.truth.Truth.assertThat
 import com.google.gson.Gson
 import com.google.gson.JsonParseException
 import com.google.gson.stream.JsonReader
-import okhttp3.Interceptor
 import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
-import okhttp3.Response
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.SocketPolicy
@@ -25,7 +24,6 @@ import okio.BufferedSink
 import okio.ByteString
 import okio.ByteString.Companion.encodeUtf8
 import okio.GzipSink
-import okio.IOException
 import okio.buffer
 import org.junit.Rule
 import org.junit.jupiter.api.assertThrows
@@ -38,25 +36,6 @@ import java.net.HttpURLConnection.HTTP_NO_CONTENT
 
 @ExtendWith(NoLoggerRule::class)
 internal class ChuckerInterceptorTest {
-    enum class ClientFactory {
-        APPLICATION {
-            override fun create(interceptor: Interceptor): OkHttpClient {
-                return OkHttpClient.Builder()
-                    .addInterceptor(interceptor)
-                    .build()
-            }
-        },
-        NETWORK {
-            override fun create(interceptor: Interceptor): OkHttpClient {
-                return OkHttpClient.Builder()
-                    .addNetworkInterceptor(interceptor)
-                    .build()
-            }
-        };
-
-        abstract fun create(interceptor: Interceptor): OkHttpClient
-    }
-
     @get:Rule val server = MockWebServer()
 
     private val serverUrl = server.url("/") // Starts server implicitly
@@ -419,7 +398,7 @@ internal class ChuckerInterceptorTest {
         server.enqueue(MockResponse())
         val client = factory.create(chuckerInterceptor)
 
-        val request = "\u0080".encodeUtf8().toRequestBody().toServerRequest()
+        val request = "\u0080".encodeUtf8().toRequestBody().toServerRequest(serverUrl)
         client.newCall(request).execute().body!!.close()
 
         val transaction = chuckerInterceptor.expectTransaction()
@@ -432,7 +411,7 @@ internal class ChuckerInterceptorTest {
         server.enqueue(MockResponse())
         val client = factory.create(chuckerInterceptor)
 
-        val request = "Hello, world!".toRequestBody().toServerRequest()
+        val request = "Hello, world!".toRequestBody().toServerRequest(serverUrl)
         client.newCall(request).execute().readByteStringBody()
         val serverRequestContent = server.takeRequest().body.readByteString()
 
@@ -445,7 +424,7 @@ internal class ChuckerInterceptorTest {
         server.enqueue(MockResponse())
         val client = factory.create(chuckerInterceptor)
 
-        val request = "Hello, world!".toRequestBody().toServerRequest()
+        val request = "Hello, world!".toRequestBody().toServerRequest(serverUrl)
         client.newCall(request).execute().readByteStringBody()
 
         val transaction = chuckerInterceptor.expectTransaction()
@@ -463,7 +442,7 @@ internal class ChuckerInterceptorTest {
         val gzippedBytes = Buffer().apply {
             GzipSink(this).buffer().use { sink -> sink.writeUtf8("Hello, world!") }
         }.readByteString()
-        val request = gzippedBytes.toRequestBody().toServerRequest()
+        val request = gzippedBytes.toRequestBody().toServerRequest(serverUrl)
             .newBuilder()
             .header("Content-Encoding", "gzip")
             .build()
@@ -485,7 +464,7 @@ internal class ChuckerInterceptorTest {
         )
         val client = factory.create(chuckerInterceptor)
 
-        val request = "!".repeat(SEGMENT_SIZE.toInt() * 10).toRequestBody().toServerRequest()
+        val request = "!".repeat(SEGMENT_SIZE.toInt() * 10).toRequestBody().toServerRequest(serverUrl)
         client.newCall(request).execute().readByteStringBody()
 
         val transaction = chuckerInterceptor.expectTransaction()
@@ -538,7 +517,7 @@ internal class ChuckerInterceptorTest {
             override fun writeTo(sink: BufferedSink) {
                 content.readAll(sink)
             }
-        }.toServerRequest()
+        }.toServerRequest(serverUrl)
 
         client.newCall(oneShotRequest).execute().readByteStringBody()
 
@@ -559,188 +538,11 @@ internal class ChuckerInterceptorTest {
             override fun writeTo(sink: BufferedSink) {
                 content.readAll(sink)
             }
-        }.toServerRequest()
+        }.toServerRequest(serverUrl)
 
         client.newCall(oneShotRequest).execute().readByteStringBody()
         val serverRequestContent = server.takeRequest().body.readByteString()
 
         assertThat(serverRequestContent.utf8()).isEqualTo("Hello, world!")
-    }
-
-    private fun RequestBody.toServerRequest() = Request.Builder().url(serverUrl).post(this).build()
-
-    @ParameterizedTest
-    @EnumSource
-    fun customBodyDecoder_doesNotChangeRequestBody(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(ReversingDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-        server.enqueue(MockResponse())
-
-        val request = "Hello, world!".toRequestBody().toServerRequest()
-        client.newCall(request).execute().readByteStringBody()
-        val serverRequestContent = server.takeRequest().body.readByteString()
-
-        assertThat(serverRequestContent.utf8()).isEqualTo("Hello, world!")
-    }
-
-    @ParameterizedTest
-    @EnumSource
-    fun customBodyDecoder_doesNotChangeResponseBody(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(ReversingDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-
-        val body = Buffer().apply { writeUtf8("Hello, world!") }
-        server.enqueue(MockResponse().setBody(body))
-        val request = Request.Builder().url(serverUrl).build()
-
-        val responseBody = client.newCall(request).execute().readByteStringBody()!!
-
-        assertThat(responseBody.utf8()).isEqualTo("Hello, world!")
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = ClientFactory::class)
-    fun customBodyDecoder_isUsedForDecoding(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(LiteralBodyDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-        val request = Request.Builder().url(serverUrl)
-            .post("Hello".toRequestBody())
-            .build()
-        server.enqueue(MockResponse().setBody("Goodbye"))
-
-        client.newCall(request).execute().readByteStringBody()
-
-        val transaction = chuckerInterceptor.expectTransaction()
-
-        assertThat(transaction.requestBody).isEqualTo("Request")
-        assertThat(transaction.responseBody).isEqualTo("Response")
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = ClientFactory::class)
-    fun bodyDecoders_areUsedInAppliedOrder(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(ReversingDecoder(), LiteralBodyDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-        val request = Request.Builder().url(serverUrl)
-            .post("Hello".toRequestBody())
-            .build()
-        server.enqueue(MockResponse().setBody("Goodbye"))
-
-        client.newCall(request).execute().readByteStringBody()
-
-        val transaction = chuckerInterceptor.expectTransaction()
-
-        assertThat(transaction.requestBody).isEqualTo("olleH")
-        assertThat(transaction.responseBody).isEqualTo("eybdooG")
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = ClientFactory::class)
-    fun nextBodyDecoder_isUsed_whenPreviousDoesNotDecode(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(NoOpDecoder(), LiteralBodyDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-        val request = Request.Builder().url(serverUrl)
-            .post("Hello".toRequestBody())
-            .build()
-        server.enqueue(MockResponse().setBody("Goodbye"))
-
-        client.newCall(request).execute().readByteStringBody()
-
-        val transaction = chuckerInterceptor.expectTransaction()
-
-        assertThat(transaction.requestBody).isEqualTo("Request")
-        assertThat(transaction.responseBody).isEqualTo("Response")
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = ClientFactory::class)
-    fun bodyDecoder_canThrowIoExceptions(factory: ClientFactory) {
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(IoThrowingDecoder(), LiteralBodyDecoder()),
-        )
-        val client = factory.create(chuckerInterceptor)
-        val request = Request.Builder().url(serverUrl)
-            .post("Hello".toRequestBody())
-            .build()
-        server.enqueue(MockResponse().setBody("Goodbye"))
-
-        client.newCall(request).execute().readByteStringBody()
-
-        val transaction = chuckerInterceptor.expectTransaction()
-
-        assertThat(transaction.requestBody).isEqualTo("Request")
-        assertThat(transaction.responseBody).isEqualTo("Response")
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = ClientFactory::class)
-    fun bodyDecoders_areAppliedLazily(factory: ClientFactory) {
-        val statefulDecoder = StatefulDecoder()
-        val chuckerInterceptor = ChuckerInterceptorDelegate(
-            cacheDirectoryProvider = { tempDir },
-            decoders = listOf(LiteralBodyDecoder(), statefulDecoder),
-        )
-        val client = factory.create(chuckerInterceptor)
-        val request = Request.Builder().url(serverUrl)
-            .post("Hello".toRequestBody())
-            .build()
-        server.enqueue(MockResponse().setBody("Goodbye"))
-
-        client.newCall(request).execute().readByteStringBody()
-
-        assertThat(statefulDecoder.didDecodeRequest).isFalse()
-        assertThat(statefulDecoder.didDecodeResponse).isFalse()
-    }
-
-    private class LiteralBodyDecoder : BodyDecoder {
-        override fun decodeRequest(request: Request, body: ByteString) = "Request"
-        override fun decodeResponse(response: Response, body: ByteString) = "Response"
-    }
-
-    private class ReversingDecoder : BodyDecoder {
-        override fun decodeRequest(request: Request, body: ByteString) = body.utf8().reversed()
-        override fun decodeResponse(response: Response, body: ByteString) = body.utf8().reversed()
-    }
-
-    private class NoOpDecoder : BodyDecoder {
-        override fun decodeRequest(request: Request, body: ByteString): String? = null
-        override fun decodeResponse(response: Response, body: ByteString): String? = null
-    }
-
-    private class IoThrowingDecoder : BodyDecoder {
-        override fun decodeRequest(request: Request, body: ByteString) = throw IOException("Request")
-        override fun decodeResponse(response: Response, body: ByteString) = throw IOException("Response")
-    }
-
-    private class StatefulDecoder : BodyDecoder {
-        var didDecodeRequest = false
-
-        override fun decodeRequest(request: Request, body: ByteString): String {
-            didDecodeRequest = true
-            return ""
-        }
-
-        var didDecodeResponse = false
-
-        override fun decodeResponse(response: Response, body: ByteString): String {
-            didDecodeResponse = true
-            return ""
-        }
     }
 }
