+++ library/src/test/java/com/chuckerteam/chucker/util/ClientFactory.kt
@@ -1,0 +1,23 @@
+package com.chuckerteam.chucker.util
+
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+
+internal enum class ClientFactory {
+    APPLICATION {
+        override fun create(interceptor: Interceptor): OkHttpClient {
+            return OkHttpClient.Builder()
+                .addInterceptor(interceptor)
+                .build()
+        }
+    },
+    NETWORK {
+        override fun create(interceptor: Interceptor): OkHttpClient {
+            return OkHttpClient.Builder()
+                .addNetworkInterceptor(interceptor)
+                .build()
+        }
+    };
+
+    abstract fun create(interceptor: Interceptor): OkHttpClient
+}
